+++ src/Renderables/Sprite.cpp
@@ -1,0 +1,10 @@
+/**
+ * @file Sprite.cpp
+ * @brief This file contains definitions for the Sprite class.
+ * 
+ * @author Willow Ciesialka
+ * @date 2023-06-07
+*/
+
+#include "Sprite.hpp"
+using GolfEngine::Sprite;
