+++ src/Rendering/RenderableVisitor.cpp
